+++ src/lib.rs
@@ -70,27 +70,62 @@
 //!         .with_console_report(true) // true means it reports to console too
 //!         .with_msg_flush_interval(2_000_000_000) // flushing interval is 2 billion nanoseconds = 2 seconds
 //!         .with_msg_buffer_size(100) // messages are flushed when there are more than 100 messages
-//!         .with_max_log_level(LogLevel::Debug)
+//!         //.with_max_log_level(LogLevel::Debug) // DEPRECATED! compile-time feature flags are recommended
 //!         .with_timezone(TimeZone::Local)
 //!         .include_unixnano(true) // include unixnano (u64) in the log as well as date and time
 //!         .with_logger_core(1) // core for logger affinity
 //!         .launch();
 //! 
 //!     flash_info_ct!(Hello::FlashLog);
+//!     // {"date":"20240915","level":"Info","message":"","offset":9,"src":"src\\logger_v2.rs:346","time":"20:34:30.684:921:877","topic":"World", "unixnano": 1741046422247135000}
 //!     flash_info_ct!(Hello::World);
+//!     // {"date":"20240915","level":"Info","message":"","offset":9,"src":"src\\logger_v2.rs:347","time":"20:34:30.684:922:238","topic":"FlashLog", "unixnano": 1741046422247135000}
 //!     flash_info_ct!("Hello");
+//!     // {"date":"20240915","level":"Info","message":"","offset":9,"src":"src\\logger_v2.rs:348","time":"20:34:30.684:922:488","topic":"Hello", "unixnano": 1741046422247135000}
 //!     flash_info_ct!("Hello"; "FlashLog");
+//!     // {"date":"20240915","level":"Info","message":"FlashLog","offset":9,"src":"src\\logger_v2.rs:349","time":"20:34:30.684:922:739","topic":"Hello", "unixnano": 1741046422247135000}
 //!     flash_info_ct!("Hello"; "FlashLog"; version = "0.1.0");
+//!     // {"data":{"version":"0.1.0"},"date":"20240915","level":"Info","message":"FlashLog","offset":9,"src":"src\\logger_v2.rs:350","time":"20:34:30.684:924:813","topic":"Hello", "unixnano": 1741046422247135000}
 //!     flash_info_ct!("Hello"; "FlashLog"; version = "0.1.0", author = "John Doe");
+//!     // {"data":{"author":"John Doe","version":"0.1.0"},"date":"20240915","level":"Info","message":"FlashLog","offset":9,"src":"src\\logger_v2.rs:351","time":"20:34:30.684:925:143","topic":"Hello", "unixnano": 1741046422247135000}
 //!     flash_info_ct!(version = "0.1.0");
+//!     // {"data":{"version":"0.1.0"},"date":"20240915","level":"Info","message":"","offset":9,"src":"src\\logger_v2.rs:352","time":"20:34:30.684:925:394","topic":"", "unixnano": 1741046422247135000}
 //!     flash_info_ct!(version = "0.1.0", author = "John Doe");
+//!     // {"data":{"author":"John Doe","version":"0.1.0"},"date":"20240915","level":"Info","message":"","offset":9,"src":"src\\logger_v2.rs:353","time":"20:34:30.684:925:654","topic":"", "unixnano": 1741046422247135000}
 //!     flash_info_ct!("topic1"; "message {} {}", 1, 2);
+//!     // {"data":"","date":"20240915","level":"Info","message":"message 1 2","offset":9,"src":"src\\logger_v2.rs:354","time":"20:34:30.684:925:955","topic":"topic1", "unixnano": 1741046422247135000}
 //!     flash_info_ct!("topic2"; "message {} {}", 1, 2; struct_info = 1, struct_info2 = 2);
+//!     // {"data":{"struct_info":1,"struct_info2":2},"date":"20240915","level":"Info","message":"message 1 2","offset":9,"src":"src\\logger_v2.rs:355","time":"20:34:30.684:926:847","topic":"topic2", "unixnano": 1741046422247135000}
 //!     flush!(); // this flushes regardless of the buffer size and flush interval
 //! 
 //!     Ok(())
 //! }
 //! ```
+//! 
+//! # Benchmark Results
+//! 
+//! Performance comparisons showing time taken per log message:
+//! 
+//! ## Test machine: Ryzen 7 7700, 3.8 Ghz
+//! | Logger    | i32           | 80 byte struct  |
+//! | --------- | ------------- | --------------- |
+//! | flashlog  | 30 ns         | 40 ns           |
+//! | ftlog     | 260 ns        | 480 ns          |
+//! | fast_log  | 410 ns        | 358 ns          |
+//! | slog      | 250 ns        | 452 ns          |
+//! | fern      | 3,813 ns      | 3,962 ns        |
+//! | tracing   | 4,003 ns      | 4,258 ns        | 
+//! 
+//! ## Test machine: i5-14400F, 2.5Ghz
+//! 
+//! | Logger    | i32           | 80 byte struct  |
+//! | --------- | ------------- | --------------- |
+//! | flashlog  | 64 ns         | 89 ns           |
+//! | ftlog     | 323 ns        | 581 ns          |
+//! | fast_log  | 500 ns        | 500 ns          |
+//! | slog      | 324 ns        | 604 ns          |
+//! | fern      | 4,732 ns      | 5,714 ns        |
+//! | tracing   | 5,177 ns      | 6,190 ns        |
 //! 
 //! For more detailed information about each module and function, please refer to the individual documentation pages.