+++ CHANGELOG.md
@@ -3,7 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [0.3.0]
+## [0.3.0] - 2025-03-14
+ - performance table in doc comments
+ - `flash_xxx_ct!` macros uses only compile-time filter
+
+## [0.3.0] - 2024-03-14
  - in the log macro, the struct data is cloned. The cloning is done in the current thread and serialized in the logging thread.
  - include_unixnano option added
  - user can choose what core to choose for affinity (with_logger_core)